- Function Pointer Experiment Seems like Zig have some kind of function overriding, but it is limited to any non-comptime type.
I am going to experiment the more complicated version
trying to override a function withing a struct.
It might be impossible, but because ultimately, I
need to convert some of the sb7 material into zig
so that I can learn how to implement openGL by
zig standard; thus, I need to understand the possibility,
of this trick.